BASE_ADDR def

+++ src/cross_lib/display/graphics_mode/memory_mapped_graphics.h
@@ -37,7 +37,7 @@
 	#define BASE_ADDR (12288+40)
 	#define COLOR_ADDR (BASE_ADDR+1024)
 #elif defined(__COMX__)
-	#define BASE_ADDR (unsigned short) 0xF800
+	#define BASE_ADDR 0xF800
 #endif
 
 #if defined(__MSX__)
